Moved commands to spi package
